requirements.txt
colorama==0.4.6
pyaml==21.10.1